+++ backend/src/test/java/com/ody/auth/domain/authorizeType/NewUserForNewDeviceTest.java
@@ -1,8 +1,8 @@
 package com.ody.auth.domain.authorizeType;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
+import com.ody.auth.domain.authpolicy.NewUserForNewDevice;
 import com.ody.mate.domain.Nickname;
 import com.ody.member.domain.DeviceToken;
 import com.ody.member.domain.Member;
